refactor: update css
