+++ src/component/UI/AddCard/AddCard.module.css
@@ -1,15 +1,21 @@
 .formInput {
   width: 100%;
-  margin: 2px 0px 15px 0;
+  margin: 2px 0px 10px 0;
   padding: 10px;
   border: 1px solid rgb(200, 148, 93);
   font-size: 18px;
 }
 
 .formCard {
-    max-width: 600px;
+  display: flex;
+  flex-direction: column;
+  max-width: 600px;
+  margin: 0 auto;
 }
 
+.formBlock {
+  margin: 5px 0 10px 0;
+}
 .label {
   font-weight: 600;
   color: #C8945D;
