+++ src/page/404/404.tsx
@@ -1,7 +1,5 @@
 import React from 'react';
 
-function Page404() {
+export function Page404() {
   return <h1>Not found cats (((</h1>;
 }
-
-export default Page404;
