--- src/page/forms/form.module.css
@@ -1,11 +1,0 @@
-.formCard {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-}
-
-.nameField {
-    justify-content: center;
-    display: flex;
-    font-size: 24px;
-}
