+++ src/page/forms/forms.module.css
@@ -1,0 +1,5 @@
+.nameField {
+    justify-content: center;
+    display: flex;
+    font-size: 24px;
+}
