+++ src/page/forms/forms.tsx
@@ -1,14 +1,12 @@
 import React from 'react';
 import { AddCard } from '../../component/UI/AddCard/AddCard';
-import classForm from './form.module.css';
+import classForm from './forms.module.css';
 
 export const Forms = () => {
   return (
     <>
       <div className={classForm.nameField}>Enter the cats details</div>
-      <div className={classForm.formCard}>
-        <AddCard />
-      </div>
+      <AddCard />
     </>
   );
 };
